+++ sdk/src/main/java/com/hedera/hashgraph/sdk/PrivateKeyED25519.java
@@ -19,6 +19,7 @@
  */
 package com.hedera.hashgraph.sdk;
 
+import com.hedera.hashgraph.sdk.Utils.Bip32Utils;
 import org.bouncycastle.asn1.ASN1OctetString;
 import org.bouncycastle.asn1.DEROctetString;
 import org.bouncycastle.asn1.pkcs.PrivateKeyInfo;
@@ -185,7 +186,7 @@ class PrivateKeyED25519 extends PrivateKey {
             throw new IllegalStateException("this private key does not support derivation");
         }
 
-        if ((index & 0x80000000) != 0) {
+        if (Bip32Utils.isHardenedIndex(index)) {
             throw new IllegalArgumentException("the index should not be pre-hardened");
         }